rasberry py pico w rundlaufkanal
This is a section of a project that monitors and controls the functionality of a Rundlaufkanal using LDA (Laser Doppler Anemometry) and PIV (Particle Image Velocimetry) measurement techniques. This project segment includes the Python application and the MicroPython programs that control the Raspberry Pi Pico W and read data from several sensors (e.g., SCT013).
SCT013  Strom messung Sensor.(leistung)